fix(toss): address code review findings on position sync PR
/ecc:review-pr(PR #29)에서 나온 6개 Important 지적 반영.

1. PositionSyncFeed.run()의 예외 처리가 스택트레이스 없이 warning만 찍고
   조용히 재시도만 함 — exc_info=True 추가, 3회 연속 실패 시
   EventType.ERROR를 발행해 대시보드에서 확인 가능하도록 함.

2. 해외주식 소수점 수량(fractional investing) 처리 시 전체 동기화가 멈추는
   버그 — _safe_int() 대신 _safe_float()로 quantity 파싱. 연쇄적으로 다음도
   int→float/Decimal 변경: BalanceItem.qty, PositionUpdate.qty,
   PositionSnapshot.qty, positions.qty(DDL, SQLite type affinity 덕에 기존
   DB도 안전), PositionItem.qty(API 응답, avg_price/book_value와 동일하게
   Decimal). 봇이 직접 내는 주문 수량(OrderItem.qty 등)은 항상 정수라
   손대지 않음.

3. core/marketdata/normalizer.py의 normalize_toss_holdings() 삭제 — 호출부
   0개, get_balance()와 완전 중복된 dead code.

4. BalanceInfo.total_eval_amount/total_profit_loss가 KRW 보유분만 합산하고
   USD는 조용히 빠지는데 타입에 그 사실이 안 드러남 — total_eval_amount_krw
   /total_profit_loss_krw로 이름 변경해 범위를 명시.

5. PositionUpdate.change: str을 Literal["opened", "updated", "closed"]로
   좁혀 오타 방지.

6. marketValue/profitLoss 내부 필드(amount/rate)가 스펙과 다르게 누락돼도
   조용히 0으로 떨어지던 것 — _safe_nested_dict/_safe_get 헬퍼로 경고 로그를
   남기도록 함 (예외는 던지지 않되 silent 하지 않게).

부가: sync_once()가 (미래의 수동 새로고침 API 등과) 동시에 호출돼도
SELECT→INSERT 사이 경쟁 상태가 없도록 asyncio.Lock으로 직렬화.

테스트 8개 추가: 소수점 수량, 중첩 필드 누락, run() 루프 일시적 실패 복구,
연속 실패 시 ERROR 이벤트 발행, 동시 호출 직렬화. 전체 667 passed.

+++ core/store/schema.py
@@ -16,7 +16,10 @@ CREATE TABLE IF NOT EXISTS positions (
     symbol      TEXT    NOT NULL,
     market      TEXT    NOT NULL CHECK(market IN ('domestic', 'overseas')),
     env         TEXT    NOT NULL DEFAULT 'prod',
-    qty         INTEGER NOT NULL DEFAULT 0,
+    -- REAL: 해외주식(미국)은 Toss에서 소수점 단위 매매(fractional investing)를
+    -- 지원해 정수가 아닐 수 있다. SQLite는 컬럼 타입과 무관하게 REAL 값을
+    -- 그대로 저장하므로(type affinity), 기존 DB에도 안전하게 적용된다.
+    qty         REAL    NOT NULL DEFAULT 0.0,
     avg_price   REAL    NOT NULL DEFAULT 0.0,
     opened_at   TEXT    NOT NULL,
     updated_at  TEXT    NOT NULL,
